Bind creator profile inputs to store
